Replace Lift tokenizer with built-in simple tokenizer
This lets us fix a couple of issues with lift's tokenizer
(though I'm sure upstream will too, posted to their list),
and lets us use Scala-native stuff like Source. The
main point, if there is any, is to just cut down dependencies,
we may as well be a standalone simple JSON library. Not like
JSON is complicated, the larger JSON libs are only large
because they have a bunch of convenience API and extra
functionality, not because of the parser/renderer.