add an AllFields object subclassing Fields, in addition to IncludedFields and ExcludedFields
This is needed for e.g. the overload of find() that takes fields and also limit,
if you just want to limit but still want all the fields.
It is redundant with Fields being an Option[Fields] in the FindOptions class,
but needed since the find() method parameters are not optional.