Add some initial hacky stuff to runtime-check getUnwrappedAs
I think the checkCast() can be done a lot better because we
know there's a fixed set of types (BValue) that can be in
a BObject. So we can use a match statement over BValue and
only then look at the manifest to cast to.

Also the unit tests on this have some holes, to be filled in.