Generate a private copy of bson-java for casbah and hammersmith
Rather than relying on the bson-java project.

This is because the mongo-java-driver we use in bson-java
may not be the same as what casbah and hammersmith pull
in.

bson-java at this point gets compiled primarily so it
can run its tests, though once mongo-java-driver has
a more predictable version maybe it will be useful to
other people.