Improve DBObject<->BObject conversion, and move to bson-java
The main thing is that as we recurse we convert
child objects and arrays into DBObject, not into
Java maps and lists. This is maybe faster, but more
importantly it seems the previous approach confused
Casbah (or the Java driver) into nesting arrays inside
an outer array. Or possibly it was our bug. Anyway,
fixed by this.

Also moved all the DBObject stuff to bson-java since it
is theoretically generic to anything using the Java driver.