massive internal refactor to replace composers with encoders/decoders
A driver (mongo-java-driver or channel) now provides a collection
that takes type classes to encode/decode the queries and results.

The public API is mostly unchanged, except that composers are
gone.

This patch takes most steps in the direction of inverting the
dependency order; the goal is that the mongo Collection types
are below rather than above the BObject stuff in the stack,
i.e. the mongo protocol layer should not have any dependency
on the particular kinds of query/entity object.

The primary remaining problem is that the encoders/decoders
are not really shared between the mongo-java-driver and the
channel backends. You have to create codecs that match
the driver. This means that the drivers have to know how
to create specific concrete codecs, which means they
have to depend on BObject etc.
This is a broken situation for sure.

One possible fix is to implement EncodeBuffer and DecodeBuffer
in some non-netty way to use with mongo-java-driver, and then
always go to/from mongo-java-driver DBObject by going through
a byte buffer (serialize/deserialize), eliminating the current
optimization which converts BObject directly to/from DBObject.
This would be slow but would fix the semantics (since all
codecs would now be based on serialization).

Another option is to more explicitly support multiple kinds
of codec and make them typesafe, so at compile time we'd
force matching the driver and the codecs. This would probably
make a big mess just to support mongo-java-driver though.